Refractor: All Controllers

+++ app/Controllers/ProductController.php
@@ -26,8 +26,14 @@ class ProductController extends Controller
   public function getAllProducts(Request $request, Response $response, array $args)
   {
     try {
-      $id_producer = $args['id'];
+      $id_producer = $args['id'] ?? null;
+
+      if (!$id_producer) {
+        throw new Exception("Tous les champs sont obligatoires", 400);
+      }
+
       $products = $this->db->product->getAllbyidproducer($id_producer);
+
       return sendJSON($response, $products, 200);
     } catch (Exception $e) {
       return sendError($response, $e->getMessage());
@@ -38,7 +44,12 @@ class ProductController extends Controller
   public function getProduct(Request $request, Response $response, array $args)
   {
     try {
-      $userId = $args['id'];
+      $userId = $args['id'] ?? null;
+
+      if (!$userId) {
+        throw new Exception("Tous les champs sont obligatoires", 400);
+      }
+
       $product = $this->db->product->getProductById($userId);
 
       return sendJSON($response, $product, 200);
@@ -53,9 +64,8 @@ class ProductController extends Controller
     try {
       $user = $request->getAttribute('user');
       $userId = $user->id;
+
       $data = $request->getParsedBody();
-
-
 
       $name = $data['name'] ?? null;
       $description = $data['description'] ?? null;
@@ -65,18 +75,13 @@ class ProductController extends Controller
       $stock = $data['stock'] ?? null;
       $image = $request->getUploadedFiles()['image'] ?? null;
 
-      if (!$image) {
-        throw new Exception("L'image est obligatoire", 400);
-      }
-
-
       if (!$name || !$description || !$type || !$price || !$unit || !$stock || !$image) {
         throw new Exception("Tous les champs sont obligatoires", 400);
       }
 
       $id = uniqid();
+      $id_producer = $this->db->producer->getProducerByUserId($userId)['id_producer'];
 
-      $id_producer = $this->db->producer->getProducerByUserId($userId)[0]['id_producer'];
       if (!$id_producer) {
         throw new Exception("Vous n'avez pas de producteur", 400);
       }
@@ -84,7 +89,6 @@ class ProductController extends Controller
       $directory = dirname(dirname(__DIR__)) . '/ressource/image';
 
       verificationImage($image);
-
       $imageName = uploadImage($image, $directory);
 
       $product = $this->db->product->addProduct($id, $name, $description, $type, $price, $unit, $stock, $imageName, $id_producer);
@@ -100,7 +104,7 @@ class ProductController extends Controller
   public function updateProduct(Request $request, Response $response, array $args)
   {
     try {
-      $productId = $args['id'];
+      $productId = $args['id'] ?? null;
       $rawdata = file_get_contents("php://input");
       parse_str($rawdata, $data);
 
@@ -112,7 +116,7 @@ class ProductController extends Controller
       $stock = $data['stock'] ?? null;
       $image = $data['image'] ?? null;
 
-      if (!$name || !$description || !$type || !$price || !$unit || !$stock) {
+      if (!$productId || !$name || !$description || !$type || !$price || !$unit || !$stock) {
         throw new Exception("Tous les champs sont obligatoires", 400);
       }
 
@@ -126,7 +130,12 @@ class ProductController extends Controller
   public function deleteProduct(Request $request, Response $response, array $args)
   {
     try {
-      $productId = $args['id'];
+      $productId = $args['id'] ?? null;
+
+      if (!$productId) {
+        throw new Exception("Tous les champs sont obligatoires", 400);
+      }
+
       $this->db->product->deleteProductById($productId);
 
       return sendJSON($response, "Le produit a bien été supprimé", 200);
